UWM-HUD: Calendar Event display improvements (#166)
* UWM-HUD: Change Calendar Event display
Remove the duration and private note fields from the display template. Add a function to make the start_date_time value available as a Datetime object in Twig so that we may print it out in the details setion. Calculate the end time based on the start time plus the duration.

* UWM-HUD: Bring webform config up to date with latest version
* UWM-HUD: Don't display 'Employee Event Categories' in the category filter
* UWM-HUD: Apply PHPCS fixes